renamed TruthTable's, Circuit's, DatabaseGenerator's (also Parameters') ints

+++ src/circuit/CircuitParameters.hpp
@@ -40,16 +40,16 @@ public:
   CircuitParameters()                                               = default;
 
   /// @todo: rewrite getters and setters
-  std::string                                        d_name         = "";
-  int                                                d_numInputs    = 0;
-  int                                                d_numConstants = 0;
-  int                                                d_numOutputs   = 0;
-  unsigned                                           d_maxLevel     = 0;
-  int                                                d_numEdges     = 0;
-  int                                                d_numGates     = 0;
-  double                                             d_size         = 0.;
-  double                                             d_area         = 0.;
-  std::string                                        d_hashCode     = "";
-  std::map<std::string, int>                         d_numElementsOfEachType;
-  std::map<std::pair<std::string, std::string>, int> d_numEdgesOfEachType;
+  std::string                                           d_name      = "";
+  uint32_t                                              d_numInputs = 0;
+  uint32_t                                              d_numConstants = 0;
+  uint32_t                                              d_numOutputs   = 0;
+  uint32_t                                              d_maxLevel     = 0;
+  uint32_t                                              d_numEdges     = 0;
+  uint32_t                                              d_numGates     = 0;
+  double                                                d_size         = 0.;
+  double                                                d_area         = 0.;
+  std::string                                           d_hashCode     = "";
+  std::map<std::string, size_t>                         d_numElementsOfEachType;
+  std::map<std::pair<std::string, std::string>, size_t> d_numEdgesOfEachType;
 };
